requirements.txt
clams-python==1.3.2
mmif-python[cv]
torch>=1.10.0
transformers>=5.0.0
torchvision>=0.15.0
Pillow>=8.0.0
accelerate
sentencepiece
tqdm
pyyaml
qwen-vl-utils
